more thorough testing of FSTransact implementation

+++ esky/fstransact.py
@@ -14,6 +14,10 @@ Currently supported platforms are:
     * Windows Vista and later, using MoveFileTransacted and friends
     * err..that's it for the moment, actually
 
+Although transactions are not supported on POSIX platforms, the way Esky
+structures its filesystem operations means that the program is always safe
+as long as you can atomically replace a file.
+
 
 """
 
@@ -39,6 +43,7 @@ if sys.platform == "win32":
             CopyFileTransacted = kernel32.CopyFileTransactedA
             DeleteFileTransacted = kernel32.DeleteFileTransactedA
             RemoveDirectoryTransacted = kernel32.RemoveDirectoryTransactedA
+            CreateDirectoryTransacted = kernel32.CreateDirectoryTransactedA
         except (WindowsError,AttributeError):
             CreateTransaction = None
 
@@ -52,6 +57,8 @@ def files_differ(file1,file2):
          return True
     if stat1.st_size != stat2.st_size:
         return True
+    assert not os.path.isdir(file1)
+    assert not os.path.isdir(file2)
     f1 = open(file1,"rb")
     try:
         f2 = open(file2,"rb")
@@ -83,42 +90,97 @@ if CreateTransaction:
             self.trnid = CreateTransaction(None,0,0,0,0,None,"")
 
         def move(self,source,target):
-            if not files_differ(source,target):
-                self.remove(source)
-                return
-            srcenc = source.encode(sys.getfilesystemencoding())
-            tgtenc = target.encode(sys.getfilesystemencoding())
+            if os.path.isdir(source):
+                if os.path.isdir(target):
+                    s_names = os.listdir(source)
+                    for nm in s_names:
+                        self.move(os.path.join(source,nm),
+                                  os.path.join(target,nm))
+                    for nm in os.listdir(target):
+                        if nm not in s_names:
+                            self.remove(os.path.join(target,nm))
+                    self.remove(source)
+                else:
+                    self._move(source,target)
+            else:
+                if os.path.isdir(target) or files_differ(source,target):
+                    self._move(source,target)
+                else:
+                    self._remove(source)
+
+        def _move(self,source,target):
+            source = source.encode(sys.getfilesystemencoding())
+            target = target.encode(sys.getfilesystemencoding())
             if os.path.exists(target):
-                MoveFileTransacted(tgtenc,tgtenc+".old",None,None,1,self.trnid)
-                MoveFileTransacted(srcenc,tgtenc,None,None,1,self.trnid)
+                target_old = target + ".old"
+                while os.path.exists(target_old):
+                    target_old += ".old"
+                MoveFileTransacted(target,target_old,None,None,1,self.trnid)
+                MoveFileTransacted(source,target,None,None,1,self.trnid)
                 try:
-                    DeleteFileTransacted(tgtenc+".old",self.trnid)
+                    self.remove(target_old)
                 except EnvironmentError:
                     pass
             else:
-                MoveFileTransacted(srcenc,tgtenc,None,None,1,self.trnid)
+                MoveFileTransacted(source,target,None,None,1,self.trnid)
 
         def copy(self,source,target):
-            if not files_differ(source,target):
-                return
-            srcenc = source.encode(sys.getfilesystemencoding())
-            tgtenc = target.encode(sys.getfilesystemencoding())
+            if os.path.isdir(source):
+                if os.path.isdir(target):
+                    s_names = os.listdir(source)
+                    for nm in s_names:
+                        self.copy(os.path.join(source,nm),
+                                  os.path.join(target,nm))
+                    for nm in os.listdir(target):
+                        if nm not in s_names:
+                            self.remove(os.path.join(target,nm))
+                else:
+                    self._copy(source,target)
+            else:
+                if os.path.isdir(target) or files_differ(source,target):
+                    self._copy(source,target)
+
+        def _copy(self,source,target):
+            source = source.encode(sys.getfilesystemencoding())
+            target = target.encode(sys.getfilesystemencoding())
             if os.path.exists(target):
-                MoveFileTransacted(tgtenc,tgtenc+".old",None,None,1,self.trnid)
-                CopyFileTransacted(srcenc,tgtenc,None,None,None,0,self.trnid)
+                target_old = target + ".old"
+                while os.path.exists(target_old):
+                    target_old += ".old"
+                MoveFileTransacted(target,target_old,None,None,1,self.trnid)
+                self._do_copy(source,target)
                 try:
-                    DeleteFileTransacted(tgtenc+".old",self.trnid)
+                    self.remove(target_old)
                 except EnvironmentError:
                     pass
             else:
-                CopyFileTransacted(srcenc,tgtenc,None,None,None,0,self.trnid)
+                target_old = None
+                if os.path.isdir(target):
+                    target_old = target + ".old"
+                    while os.path.exists(target_old):
+                        target_old += ".old"
+                    MoveFileTransacted(target,target_old,None,None,1,self.trnid)
+                self._do_copy(source,target)
+                if target_old is not None:
+                    self.remove(target_old)
+
+        def _do_copy(self,source,target):
+            if os.path.isdir(source):
+                CreateDirectoryTransacted(None,target,0,self.trnid)
+                for nm in os.listdir(source):
+                    self._do_copy(os.path.join(source,nm),
+                                  os.path.join(target,nm))
+            else:
+                CopyFileTransacted(source,target,None,None,None,0,self.trnid)
 
         def remove(self,target):
-            tgtenc = target.encode(sys.getfilesystemencoding())
+            target = target.encode(sys.getfilesystemencoding())
             if os.path.isdir(target):
-                RemoveDirectoryTransacted(tgtenc,self.trnid)
+                for nm in os.listdir(target):
+                    self.remove(os.path.join(target,nm))
+                RemoveDirectoryTransacted(target,self.trnid)
             else:
-                DeleteFileTransacted(tgtenc,self.trnid)
+                DeleteFileTransacted(target,self.trnid)
 
         def commit(self):
             CommitTransaction(self.trnid)
@@ -139,10 +201,23 @@ else:
             self.pending = []
 
         def move(self,source,target):
-            if files_differ(source,target):
-                self.pending.append(("_move",source,target))
+            if os.path.isdir(source):
+                if os.path.isdir(target):
+                    s_names = os.listdir(source)
+                    for nm in s_names:
+                        self.move(os.path.join(source,nm),
+                                  os.path.join(target,nm))
+                    for nm in os.listdir(target):
+                        if nm not in s_names:
+                            self.remove(os.path.join(target,nm))
+                    self.remove(source)
+                else:
+                    self.pending.append(("_move",source,target))
             else:
-                self.pending.append(("_remove",source))
+                if os.path.isdir(target) or files_differ(source,target):
+                    self.pending.append(("_move",source,target))
+                else:
+                    self.pending.append(("_remove",source))
 
         def _move(self,source,target):
             if sys.platform == "win32" and os.path.exists(target):
@@ -160,15 +235,40 @@ else:
                     raise
                 else:
                     try:
-                        os.unlink(target_old)
+                        self._remove(target_old)
                     except EnvironmentError:
                         pass
             else:
+                target_old = None
+                if os.path.isdir(target) and os.path.isfile(source):
+                    target_old = target + ".old"
+                    while os.path.exists(target_old):
+                        target_old = target_old + ".old"
+                    os.rename(target,target_old)
+                elif os.path.isfile(target) and os.path.isdir(source):
+                    target_old = target + ".old"
+                    while os.path.exists(target_old):
+                        target_old = target_old + ".old"
+                    os.rename(target,target_old)
                 os.rename(source,target)
+                if target_old is not None:
+                    self._remove(target_old)
 
         def copy(self,source,target):
-            if files_differ(source,target):
-                self.pending.append(("_copy",source,target))
+            if os.path.isdir(source):
+                if os.path.isdir(target):
+                    s_names = os.listdir(source)
+                    for nm in s_names:
+                        self.copy(os.path.join(source,nm),
+                                  os.path.join(target,nm))
+                    for nm in os.listdir(target):
+                        if nm not in s_names:
+                            self.remove(os.path.join(target,nm))
+                else:
+                    self.pending.append(("_copy",source,target))
+            else:
+                if os.path.isdir(target) or files_differ(source,target):
+                    self.pending.append(("_copy",source,target))
 
         def _copy(self,source,target):
             if sys.platform == "win32" and os.path.exists(target):
@@ -177,7 +277,7 @@ else:
                     target_old = target_old + ".old"
                 os.rename(target,target_old)
                 try:
-                    shutil.copy2(source,target)
+                    self._do_copy(source,target)
                 except:
                     os.rename(target_old,target)
                     raise
@@ -187,7 +287,26 @@ else:
                     except EnvironmentError:
                         pass
             else:
+                target_old = None
+                if os.path.isdir(target) and os.path.isfile(source):
+                    target_old = target + ".old"
+                    while os.path.exists(target_old):
+                        target_old = target_old + ".old"
+                    os.rename(target,target_old)
+                elif os.path.isfile(target) and os.path.isdir(source):
+                    target_old = target + ".old"
+                    while os.path.exists(target_old):
+                        target_old = target_old + ".old"
+                    os.rename(target,target_old)
+                self._do_copy(source,target)
+                if target_old is not None:
+                    self._remove(target_old)
+
+        def _do_copy(self,source,target):
+            if os.path.isfile(source):
                 shutil.copy2(source,target)
+            else:
+                shutil.copytree(source,target)
 
         def remove(self,target):
             self.pending.append(("_remove",target))
@@ -195,7 +314,9 @@ else:
         def _remove(self,target):
             if os.path.isfile(target):
                 os.unlink(target)
-            else:
+            elif os.path.isdir(target):
+                for nm in os.listdir(target):
+                    self._remove(os.path.join(target,nm))
                 os.rmdir(target)
 
         def commit(self):
